requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0